+++ Plugins/StatusEverywhere/StatusEverywhere.plugin.js
@@ -8,24 +8,40 @@ var StatusEverywhere = (() => {
 			catch(err) {reject(err);}
 		});
 	});
-	const config = {"info":{"name":"StatusEverywhere","authors":[{"name":"Zerebos","discord_id":"249746236008169473","github_username":"rauenzi","twitter_username":"ZackRauen"}],"version":"0.4.0","description":"Adds user status everywhere Discord doesn't. Support Server: bit.ly/ZeresServer","github":"https://github.com/rauenzi/BetterDiscordAddons/tree/master/Plugins/StatusEverywhere","github_raw":"https://raw.githubusercontent.com/rauenzi/BetterDiscordAddons/master/Plugins/StatusEverywhere/StatusEverywhere.plugin.js"},"changelog":[{"title":"Bugs Squashed","type":"fixed","items":["Updated for Discord changes."]}],"main":"index.js"};
+	const config = {"info":{"name":"StatusEverywhere","authors":[{"name":"Zerebos","discord_id":"249746236008169473","github_username":"rauenzi","twitter_username":"ZackRauen"}],"version":"0.4.1","description":"Adds user status everywhere Discord doesn't. Support Server: bit.ly/ZeresServer","github":"https://github.com/rauenzi/BetterDiscordAddons/tree/master/Plugins/StatusEverywhere","github_raw":"https://raw.githubusercontent.com/rauenzi/BetterDiscordAddons/master/Plugins/StatusEverywhere/StatusEverywhere.plugin.js"},"changelog":[{"title":"Improvements","type":"improved","items":["Tapped into Discord's secondary status cache to show less members as offline when they aren't."]}],"main":"index.js"};
 	const compilePlugin = ([Plugin, Api]) => {
 		const plugin = (Plugin, Api) => {
     const {Patcher, WebpackModules, DiscordModules} = Api;
-    return class HideDisabledEmojis extends Plugin {
+    return class StatusEverywhere extends Plugin {
         onStart() {
+			this.showAnnouncement();
             const Avatar = WebpackModules.getByProps("AvatarWrapper");
             const original = Avatar.default;
             Patcher.before(Avatar, "default", (_, args) => {
                 if (args[0].status) return;
                 const id = args[0].src.split("/")[4];
-                args[0].status = DiscordModules.UserStatusStore.getStatus(id);
+                const cached = DiscordModules.UserStatusStore.getStatus(id);
+				args[0].status = cached == "offline" ? DiscordModules.UserStatusStore.getStatus(id, DiscordModules.SelectedGuildStore.getGuildId()) : cached;
             });
             Object.assign(Avatar.default, original);
         }
         
         onStop() {
             Patcher.unpatchAll();
+        }
+		
+		showAnnouncement() {
+            if (window.ZeresPluginLibrary) return; // they already have it
+            const hasShownAnnouncement = PluginUtilities.loadData(this.getName(), "announcements", {localLibNotice: false}).localLibNotice;
+            if (hasShownAnnouncement) return;
+            Modals.showConfirmationModal("Local Library Notice", DiscordModules.React.createElement("span", null, `This version of ${this.getName()} is the final version that will be released using a remotely loaded library. Future versions will require my local library that gets placed in the plugins folder.`, DiscordModules.React.createElement("br"), DiscordModules.React.createElement("br"), "You can download the library now to be prepared, or wait until the next version which will prompt you to download it."), {
+                confirmText: "Download Now",
+                cancelText: "Wait",
+                onConfirm: () => {
+                    require("electron").shell.openExternal("https://betterdiscord.net/ghdl?url=https://raw.githubusercontent.com/rauenzi/BDPluginLibrary/master/release/0PluginLibrary.plugin.js");
+                }
+            });
+            PluginUtilities.saveData(this.getName(), "announcements", {localLibNotice: true});
         }
 
     };
@@ -35,7 +51,7 @@ var StatusEverywhere = (() => {
 	
 	return !global.ZLibrary ? class {
 		getName() {return config.info.name.replace(" ", "");} getAuthor() {return config.info.authors.map(a => a.name).join(", ");} getDescription() {return config.info.description;} getVersion() {return config.info.version;} stop() {}
-		showAlert() {window.mainCore.alert("Loading Error",`Something went wrong trying to load the library for the plugin. Try reloading?`);}
+		showAlert() {window.BdApi.alert("Loading Error",`Something went wrong trying to load the library for the plugin. You can try using a local copy of the library to fix this.<br /><br /><a href="https://betterdiscord.net/ghdl?url=https://raw.githubusercontent.com/rauenzi/BDPluginLibrary/master/release/0PluginLibrary.plugin.js" target="_blank">Click here to download the library!</a>`);}
 		async load() {
 			try {await global.ZLibraryPromise;}
 			catch(err) {return this.showAlert();}